pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "First fastapi project"
authors = ["Дмитрий <alexdem169@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.13.2"
annotated-types = "0.7.0"
anyio = "4.4.0"
bcrypt = "4.2.0"
certifi = "2024.7.4"
click = "8.1.7"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi = "0.112.2"
fastapi-cli = "0.0.5"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.2"
idna = "3.8"
jinja2 = "3.1.4"
mako = "1.3.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
passlib = "1.7.4"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyjwt = "2.9.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.2"
rich = "13.8.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlalchemy = "2.0.32"
starlette = "0.38.2"
typer = "0.12.5"
typing-extensions = "4.12.2"
uvicorn = "0.30.6"
uvloop = "0.20.0"
watchfiles = "0.24.0"
websockets = "13.0.1"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
